django-backend/attendance/serializers.py
from rest_framework import serializers
from attendance.models import AttendanceRecord

class AttendanceRecordSerializer(serializers.ModelSerializer):
    class Meta:
        model = AttendanceRecord
        fields = '__all__'
